Changed title and fixed formatting of h3 id
Changed "Essays" to "Essays and projects". Also fixed the format of the relevant id. Kasun mentioned that they shouldn't have capitals or spaces, so I've changed the id to essays_and_projects. 

Kasun also mentioned that in my case such ids may not be needed, but I'll leave them be for now in case removing them will affect the styling.